requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
filelock==3.13.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==1.35.0
networkx==3.3
nltk==3.9.1
numpy==1.24.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.28.1
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.30.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.13
watchdog==6.0.0
